refactor(console): agent controller (wave 4)
Moves the agent cluster out of `ChatScreen` into
`Console_Modules/agent.py` (`ConsoleAgentController`), built in
`wiring.py` alongside the other six: the lazily-built ConsoleAgentBridge
handle, the Agent rail's text derivation, the sub-agent drill-in cycle,
the "View full log" target/probe/loader, the fleet auto-open override,
the [N Sub-Agents] badge-count cache, and resume-time TOOL-marker
re-derivation.

Thirteen bodies move byte-for-byte. Four deviations, each disclosed in
the module docstring: `@work(thread=True)` cannot decorate a controller
(Textual asserts DOMNode), so the loader is dispatched by its caller
through `run_worker` with the same thread/group/name; two `self.app.X`
reaches became this controller's own framework-service properties; and
the two module-level constants `_console_subagent_counts_refresh_needed`
reads are imported in-body, because a module-level import of them would
be a real cycle (`chat_screen` imports `wiring` at line 89 and defines
them at 587).

`_sync_console_agent_section` is split rather than moved: controllers own
no DOM and its second half is nine `query_one` calls, so only the payload
derivation moved (`_console_agent_section_payload`); the equality guard,
its memo, and the writes stay on the screen. `_console_agent_runtime_
enabled` is excluded as a name-only match -- a config read whose only
consumers are the screen's chat-controller wiring, and whose identical
twin two lines below it is not name-matched.

Seven delegations and four read-write proxies stay on the screen for
consumers outside the cluster; the dead `ConsoleRunLogModal` import is
dropped after confirming no test reaches it through the module alias.
